backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pydantic==2.9.2
pydantic-settings==2.6.0
python-multipart==0.0.12
pandas==2.2.3
mlxtend==0.23.1
google-generativeai>=0.8.0
pytest==8.3.3
httpx==0.27.2
